+++ guis7-reflex/src/GUIs/Cells/Types.hs
@@ -1,21 +1,22 @@
 {-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE UnicodeSyntax #-}
+
 module GUIs.Cells.Types where
 
 import           Data.Decimal
-import           Data.Monoid
 import           Data.Text (Text)
 import qualified Data.Text as T
 
 data Size = Size
-  { width  :: Int
-  , height :: Int
+  { width  ∷ Int
+  , height ∷ Int
   } deriving (Show)
 
 data Coords
   = Coords Int Int -- col row
   deriving (Show, Eq, Ord)
 
-inBounds :: Size -> Coords -> Bool
+inBounds ∷ Size → Coords → Bool
 inBounds (Size w h) (Coords i j) = i < w && j < h && i >= 0 && j >= 0
 
 -- Expressions
@@ -38,7 +39,7 @@ data Expr
     | ENumber Decimal
     deriving (Show)
 
-getExprDeps :: Expr -> [Coords]
+getExprDeps ∷ Expr → [Coords]
 getExprDeps (ERef ref) = [ref]
 getExprDeps (EBinOp _ l r) = getExprDeps l ++ getExprDeps r
 getExprDeps (EUnOp _ sub) = getExprDeps sub
@@ -65,13 +66,14 @@ data EvalError
     | DivByZero
     deriving (Show)
 
-embedEvalResult :: EvalResult -> CellResult
+embedEvalResult ∷ EvalResult → CellResult
 embedEvalResult (Left err) = Left $ EvalError err
 embedEvalResult (Right x) = Right $ Number x
 
-showCellResult :: CellResult -> Text
+showCellResult ∷ CellResult → Text
 showCellResult (Left (EvalError (RefNotFound (Coords i j)))) = "{" <> (T.pack . show) i <> "," <> (T.pack . show) j <> "} not defined"
 showCellResult (Left (EvalError DivByZero)) = "Division by zero"
 showCellResult (Left (ParseError _)) = "Cannot parse expression"
 showCellResult (Right (Number x)) = "=" <> (T.pack . show) x
 showCellResult (Right Empty) = "(empty)"
+